lab2/wall_service/requirements.txt
fastapi
python-jose
uvicorn
numpy
pydantic
bcrypt==4.0.1
